+++ saleor/graphql/shipping/tests/mutations/test_shipping_zone_update.py
@@ -83,6 +83,7 @@ SHIPPING_ZONE_UPDATE_DEFAULT_TRUE_MUTATION = """
 def test_update_shipping_zone(
     staff_api_client, shipping_zone, permission_manage_shipping
 ):
+    # given
     name = "Parabolic name"
     description = "Description of a shipping zone."
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
@@ -92,11 +93,15 @@ def test_update_shipping_zone(
         "countries": [],
         "description": description,
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -162,6 +167,7 @@ def test_update_shipping_zone_trigger_webhook(
 def test_update_shipping_zone_default_exists(
     staff_api_client, shipping_zone, permission_manage_shipping
 ):
+    # given
     default_zone = shipping_zone
     default_zone.default = True
     default_zone.pk = None
@@ -170,11 +176,15 @@ def test_update_shipping_zone_default_exists(
 
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     variables = {"id": shipping_id, "name": "Name", "countries": [], "default": True}
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert data["errors"][0]["field"] == "default"
@@ -187,6 +197,7 @@ def test_update_shipping_zone_add_warehouses(
     warehouses,
     permission_manage_shipping,
 ):
+    # given
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     warehouse_ids = [
         graphene.Node.to_global_id("Warehouse", warehouse.pk)
@@ -199,11 +210,15 @@ def test_update_shipping_zone_add_warehouses(
         "name": shipping_zone.name,
         "addWarehouses": warehouse_ids,
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -220,6 +235,7 @@ def test_update_shipping_zone_add_second_warehouses(
     warehouse_no_shipping_zone,
     permission_manage_shipping,
 ):
+    # given
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     warehouse_id = graphene.Node.to_global_id(
         "Warehouse", warehouse_no_shipping_zone.pk
@@ -229,11 +245,15 @@ def test_update_shipping_zone_add_second_warehouses(
         "name": shipping_zone.name,
         "addWarehouses": [warehouse_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -248,6 +268,7 @@ def test_update_shipping_zone_remove_warehouses(
     warehouse,
     permission_manage_shipping,
 ):
+    # given
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
     variables = {
@@ -255,11 +276,15 @@ def test_update_shipping_zone_remove_warehouses(
         "name": shipping_zone.name,
         "removeWarehouses": [warehouse_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -273,6 +298,7 @@ def test_update_shipping_zone_remove_one_warehouses(
     warehouses,
     permission_manage_shipping,
 ):
+    # given
     for warehouse in warehouses:
         warehouse.shipping_zones.add(shipping_zone)
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
@@ -282,11 +308,15 @@ def test_update_shipping_zone_remove_one_warehouses(
         "name": shipping_zone.name,
         "removeWarehouses": [warehouse_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -302,6 +332,7 @@ def test_update_shipping_zone_replace_warehouse(
     warehouse_no_shipping_zone,
     permission_manage_shipping,
 ):
+    # given
     assert shipping_zone.warehouses.first() == warehouse
 
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
@@ -315,11 +346,15 @@ def test_update_shipping_zone_replace_warehouse(
         "addWarehouses": [add_warehouse_id],
         "removeWarehouses": [remove_warehouse_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -334,6 +369,7 @@ def test_update_shipping_zone_same_warehouse_id_in_add_and_remove(
     warehouse,
     permission_manage_shipping,
 ):
+    # given
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
     variables = {
@@ -342,11 +378,15 @@ def test_update_shipping_zone_same_warehouse_id_in_add_and_remove(
         "addWarehouses": [warehouse_id],
         "removeWarehouses": [warehouse_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert data["errors"]
@@ -362,6 +402,7 @@ def test_update_shipping_zone_add_channels(
     channel_PLN,
     permission_manage_shipping,
 ):
+    # given
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     channel_ids = [
         graphene.Node.to_global_id("Channel", channel.pk)
@@ -373,11 +414,15 @@ def test_update_shipping_zone_add_channels(
         "name": shipping_zone.name,
         "addChannels": channel_ids,
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -387,7 +432,7 @@ def test_update_shipping_zone_add_channels(
 
 
 @mock.patch(
-    "saleor.graphql.shipping.mutations.shippings."
+    "saleor.graphql.shipping.mutations.base."
     "drop_invalid_shipping_methods_relations_for_given_channels.delay"
 )
 def test_update_shipping_zone_remove_channels(
@@ -398,6 +443,7 @@ def test_update_shipping_zone_remove_channels(
     channel_PLN,
     permission_manage_shipping,
 ):
+    # given
     shipping_zone.channels.add(channel_USD, channel_PLN)
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     channel_id = graphene.Node.to_global_id("Channel", channel_USD.pk)
@@ -415,11 +461,15 @@ def test_update_shipping_zone_remove_channels(
         "name": shipping_zone.name,
         "removeChannels": [channel_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -443,6 +493,7 @@ def test_update_shipping_zone_add_and_remove_channels(
     channel_PLN,
     permission_manage_shipping,
 ):
+    # given
     shipping_zone.channels.add(channel_USD)
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     add_channel_id = graphene.Node.to_global_id("Channel", channel_PLN.pk)
@@ -454,11 +505,15 @@ def test_update_shipping_zone_add_and_remove_channels(
         "removeChannels": [remove_channel_id],
         "addChannels": [add_channel_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -474,6 +529,7 @@ def test_update_shipping_zone_same_channel_id_in_add_and_remove_list(
     channel_PLN,
     permission_manage_shipping,
 ):
+    # given
     shipping_zone.channels.add(channel_USD)
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     add_channel_id = graphene.Node.to_global_id("Channel", channel_PLN.pk)
@@ -485,11 +541,15 @@ def test_update_shipping_zone_same_channel_id_in_add_and_remove_list(
         "removeChannels": [remove_channel_id],
         "addChannels": [add_channel_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert not data["errors"]
@@ -510,6 +570,7 @@ def test_update_shipping_zone_add_invalid_warehouses(
 ):
     """Ensure an error is raised when the warehouse that has not a common
     channel with shipping zone is added."""
+    # given
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
     # warehouse with common USD channel
     warehouse_usd_id = graphene.Node.to_global_id("Warehouse", warehouses[0].pk)
@@ -533,11 +594,15 @@ def test_update_shipping_zone_add_invalid_warehouses(
         "addWarehouses": [warehouse_usd_id, warehouse_jpy_id, warehouse_pln_id],
         "addChannels": [channel_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert len(data["errors"]) == 1
@@ -555,6 +620,7 @@ def test_update_shipping_zone_add_warehouse_without_any_channel(
 ):
     """Ensure an error is raised when the warehouse that has not a common
     channel with shipping zone is added."""
+    # given
     shipping_id = graphene.Node.to_global_id("ShippingZone", shipping_zone.pk)
 
     warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
@@ -567,11 +633,15 @@ def test_update_shipping_zone_add_warehouse_without_any_channel(
         "addWarehouses": [warehouse_id],
         "addChannels": [channel_id],
     }
+
+    # when
     response = staff_api_client.post_graphql(
         UPDATE_SHIPPING_ZONE_MUTATION,
         variables,
         permissions=[permission_manage_shipping],
     )
+
+    # then
     content = get_graphql_content(response)
     data = content["data"]["shippingZoneUpdate"]
     assert len(data["errors"]) == 1